Move reactQuery.ts from store/ -> db/
