+++ packages/shared/src/db/index.ts
@@ -20,3 +20,4 @@ export { setClient } from './client';
 export type { AnySqliteDatabase } from './client';
 export * from './changeListener';
 export * from './writeObservers';
+export * from './reactQuery';
