+++ packages/shared/src/db/query.ts
@@ -3,9 +3,9 @@ import { sql } from 'drizzle-orm';
 import { createDevLogger, escapeLog, listDebugLabel } from '../debug';
 import { AnalyticsEvent } from '../domain';
 import { startTrace } from '../perf';
-import { queryClient } from '../store/reactQuery';
 import * as changeListener from './changeListener';
 import { AnySqliteDatabase, AnySqliteTransaction, client } from './client';
+import { queryClient } from './reactQuery';
 import { TableName } from './types';
 
 const logger = createDevLogger('query', false);
