+++ packages/shared/src/store/index.ts
@@ -1,5 +1,4 @@
 export * from './dbHooks';
-export * from './reactQuery';
 export * from './sync';
 export * from './useChannelPosts';
 export * from './useChannelSearch';
