+++ packages/shared/src/store/useActivityFetchers.ts
@@ -3,9 +3,9 @@ import * as api from '@tloncorp/api';
 import { useMemo } from 'react';
 
 import * as db from '../db';
+import { queryClient } from '../db/reactQuery';
 import { createDevLogger } from '../debug';
 import * as logic from '../logic';
-import { queryClient } from './reactQuery';
 import * as sync from './sync';
 
 const logger = createDevLogger('useInfiniteBucketedActivity', false);
