+++ packages/shared/src/store/useChannelPosts.ts
@@ -11,7 +11,6 @@ import { createDevLogger } from '../debug';
 import { AnalyticsEvent } from '../domain';
 import { useLiveRef, useOptimizedQueryResults } from '../logic/utilHooks';
 import { usePendingPostsInChannel } from './dbHooks';
-import { queryClient } from './reactQuery';
 import { useCurrentSession } from './session';
 import * as sync from './sync';
 import { SyncPriority } from './syncQueue';
@@ -43,7 +42,7 @@ type UseChannelPostsParams = UseChannelPostsPageParams & {
 };
 
 export const clearChannelPostsQueries = () => {
-  queryClient.invalidateQueries({ queryKey: ['channelPosts'] });
+  db.queryClient.invalidateQueries({ queryKey: ['channelPosts'] });
 };
 
 export const useChannelPosts = (options: UseChannelPostsParams) => {
